pyproject.toml
[tool.poetry]
name = "fasthtml_extn"
version = "0.1.0"
description = "An fasthtml extension framework to build webapps using python."
authors = ["Jose Antonio Dominguez <chamot11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-fasthtml = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
